p6doc -> rakudoc: Use rakudoc, .rakudoc, RAKUDOC, etc.
Summary: this is a rename commit. Everything to do with Perl 6 -> Raku
is handled here. Further commits will be needed to get the code working
with the current Documentable, and to update documentation and CI builds.

It is mainly just text substitutions, such as

- p6doc => rakudoc
- P6DOC => RAKUDOC
- perl6 => raku
- .pod6 => .rakudoc (where appropriate)
- .pm6 => .rakumod
- Perl6::Documentable => Documentable
- github:perl6/doc => github:Raku/doc.
- github:noisegul/p6doc => github:Raku/rakudoc

Minimal (untested) code changes are made as needed to handle both Raku
and Perl 6 names (e.g., RAKUDOC_TEST and P6DOC_TEST, .rakudoc and .pod6,
and so forth).

A constant `DOC-SUFFIX = '.pod6'` is introduced to facilitate changing
to `'.rakudoc'` in Raku/doc.

A few egregiously outdated items are removed from README.