Remove dead code for Index creation & usage
The index path logic is modified slightly. It now looks for an existing
rakudoc-index.json (in ~/.cache/raku, ~/.raku, ~/.perl6, ./), and if it
can't find it then will use the first of those dirs, creating the dir if
needed.

Also, beef up MAIN testing: check the output, not just that it returns
OK.

Otherwise, these changes are primarily stylistic and removing some minor
errors.